warn if there are no sample in the OTU table
thanks to Joel Hjärne Kokk for reporting this case

By design, mumu accepts empty input files. The goal is to make it as
permissive as possible, so it integrates nicely in unsupervised
pipelines. The downside is that it is hard to help users understand
what's wrong with their input data. From now on, mumu will at least
indicate that something is wrong with the OTU table.